requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pyyaml>=6.0
chromadb>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
